fix: some better conditionals in pr action (#2128)
Some steps in the build job should only be run if one of the previous
jobs has been run. Instead of encoding that specifically I'd copied over
the `if` statement from the previous job. This is kinda fragile because
if you change one condition you have to remember to change the other,
which obviously is going to be forgotten and lead to problems.