feat: run federation-audit-tests on ci (#2153)
Added the federation-audit-tests into their own workflow. These will not
fail the build, and will just post a comment summarising the current
state of the audit tests.

I wanted to output a non-failing check with more details but I couldn't
find a way to make it non-failing, and having that red cross on every
single build seemed annoying.