fix: the mysterious-external audit tests
Our implementation of the audit tests and the original one from the
guild disagreed slightly: we had 2 more failures than they did.

This turned out to be the mysterious-external tests: their test data
contains `price: 100` but the engine is returning `price: 100.0`.
`serde_json` considers these to not be equal, so we were getting a
false failure.

I've fixed that in this PR by just converting all numbers to floats
before comparing.  There may be some edge cases for which this doesn't
work, but it does give us the same numbers as the guild so I think we're
good for now.