graphql-federated-graph: introduce TypeDefinition
This is a refactoring I have had on my mind for quite some time. It
should enable simpler code in from_sdl(), rendering, as well as
composition eventually.

The main idea is that we need a Definition or TypeDefinition entity,
like in other parts of the codebase (graphql-schema-diff for example).
That definition can hold a name, a directive and a description. For
scalars, enums and union, that would be the whole of the Scalar, Enum or
Union struct, so these can eventually disappear. That will be more
concise and enable a lot more code reuse between different kinds of
definitions.

This PR is only a very partial first step, because I want to get
feedback before going further.

The other important aspect of this PR is the federated_graph::v4::view
module. This is an abstraction I have been thinking about for a while as
an alternative to the walker pattern. In the long term, we should either
keep this pattern if it is nice to work with, or switch to the same
approach to code generation for schema walkers as engine-v2. In the
short term, this is already an improvement on the status quo.