fix: propagate the correct span to subgraphs (#2243)
I've been using traces to investigate an issue today, and had enabled
propagation to see whether it was us or subgraphs that were slow. But in
my tracing view the subgraph HTTP span was marked as a sibling of our
corresponding HTTP span rather than as a child.

This fixes that by propagating the http span to children rather than
just using the current span.