composition: ignore `@composeDirective(name: "@authorized")` (#2251)
This fixes composed schemas with subgraphs schemas introspected from
async-graphql having `@authorized` twice on fields.

The `@authorized` directive is a special case. Directives used in
subgraph schemas are either built-in federation directives (`@requires`,
`@key`, etc.) or custom, composed directives with `@composeDirective`.
Since `@authorized` is not part of the federation spec, some frameworks
like async-graphql (Rust) will produce an `@composeDirective` with the
`@authorized` directive. We should not consider `@authorized` as a
composed directive however, because that means we would emit it again.

closes GB-8706