feat: initial message signing implementation (#2267)
This PR implements the majority of message signing. We now take the
config, and apply a signature to any headers that the user has requested
we sign. We do not yet support any derived components or signature
parameters or remove the hop-by-hop headers - those are still to be
done.

Fixes GB-7685
Fixes GB-7787
Fixes GB-7864