feat: signature parameters for message signing (#2280)
When implementing this I realised that the configuration I'd specified
for message signing probably didn't make sense:

1. I'm not sure there's ever any harm in including a created date -
httpsig does it by default, and if we removed it that could cause a
panic if the user provided expiry. Additionally its recommended that you
include it in the spec. So unless someone requests that we remove I
think it makes sense to just always have it
2. If we have a KID I think we should just include it, same with the
algorithm.

This just leaves the nonce parameter, which I've added support for.

Fixes GB-7858