fix: dont omit unused enums in composition (#2304)
While testing a separate composition change I wrote a schema that didn't
use an enum. I was surprised to find that enum wasn't in the output of
composition.

It seems like this was intentional, but have been speaking to @tomhoule
and we agreed it's probably not right. It shouldn't _hurt_ to remove
these types since they're not used, but we don't treat any other unused
types like this so it's a bit surprising.