chore: use cynic-parser in composition (#2307)
When working on #2305 I found that we need to parse some directives in
composition and also in federated-graph. Before I add the `@listSize`
directive I'd quite like to consolidate this. But at the moment this is
hard, because composition uses `async_graphql_parser`, whereas
federated-graph has been updated to use `cynic-parser`.

Now seems as good a time as any to change that, so this PR updates
composition to also use `cynic-parser`.

I also fixed an issue around errors parsing various `FieldSet` arguments
- the error messages were hardcoded for the fields argument on `@key`
but the function itself is now used for a few different arguments.