feat: handle the `@listSize` directive in composition (#2338)
As per the title of the PR. Also made a couple of tweaks while I was in
here:

1. Removed trailing commas from lists when rendering SDL. They didn't
hurt, but I thought they looked kind of bad? Feel free to disagree
though.
2. Changed composition tests & composition in the CLI to print a report
on GraphQL parse failures, instead of just a message. Definitely useful
in tests, but debatable whether we always want to pay the price of
building this. Comments welcome. Looks like this:

![CleanShot 2024-11-12 at 15 04
02@2x](https://github.com/user-attachments/assets/48ce647d-5379-42a6-b174-2ffd7ac23be8)

Fixes GB-7642