fix: redo `@listSize` in composition (#2346)
I added support for `@listSize` in #2338 - but once I started trying to
use it in engine I realised that I'd done a bad job. I'd kept the two
`[String]` arguments as strings - but these refer to fields & arguments
so would be better converted into the corresponding Ids so that the
engine doesn't have to do a bunch of string lookups when it's trying to
calculate the complexity.

This was harder than it sounds, because the `sizedFields` argument
refers to the fields of another object, which isn't necessarily
available at the point where we're handling most of the directives. Got
there in the end though.

Fixes GB-7642 (again)