feat: complexity cost calculation (#2356)
This PR implements _most_ of the cost calculation for complexity
control. It's relatively simple conceptually: we walk the query,
assigning a cost to each of the types & fields that are present. We
either use a default cost or a cost configured in the schema. Fields
inside lists are multiplied by the expected length of the list.

Where it gets slightly more complicated is integrating it with the
engine as it stands:

- I'd hoped to do this _before_ planning, but the engine is currently
set up such that you can't use variables until after planning has
happened. I looked into working around this but it'd either make
complexity control much more complicated or require a bit of rework
around caching. Just going to see how we get on with this approach for
now and can rework later if we need to.
- Input values aren't set up to be easy to use with anything other than
`Deserialize` impls at the moment. For calculating the cost of input
values we kind of need to walk the input value structure, but that is
very painful right now. I'm going to handle that in the next PR instead.
- I also found a bunch of dead code, which I've removed in this PR.

Fixes GB-7643
Fixed GB-7645