feat: warn on unknown directive in composition (#2362)
This change is up for debate, but it would have saved me an hour or two
of digging and seems like something a customer could run into.

I was trying to integration test `@cost`, but for some reason none of
the validations were failing. This turned out to be because the `@link`
directive was used in my schema but didn't explicitly list `@cost`. In
this scenario we still support `@cost` but we expect it to be named
something like `@federated__cost` - if we see `@cost` it is just
ignored.

The behaviour around renames is correct, but we could make it a lot
easier to notice when this sort of thing has happpened. This PR is an
attempt to do that: when we encounter a directive that we do not know
about we'll put out a warning telling the user.