fix(engine): Support __typename alias & rework key generation
It's really an edge case, but I do rely on `__typename` being available
and used for the __typename in subgraph queries. It's a bit tricky to
synchronize properly what the shapes expect and the query we generate,
the easiest that came to mind right now is to rename fields before we generate the shapes,
ensuring `__typename` is available.

So we re-use the same new key generation technique used for extra fields
& field with distinct types in their subgraph. I refactored a bit the
implementation to be common this time.