fix(engine): mutation test
There are two ordering problems with the query partitions generated by
the query-solver:

- mutation ordering is ignored
- dependencies: a -> b -> c, with a & c coming from the subgraph.

Both have dedicated functions and I was running the latter the same way
for mutations and queries. So the second would change whatever the first
one did for mutations. I just go lucky in my tests that it worked.

I also realized I had the inverted ordering of mutations... as I didn't
use the usual arrow shape for requirements.