fix: include argument values in complexity cost (#2393)
This includes the cost of argument values in complexity calculations.  

Currently we do this by serialising argument values into a
`serde_json::Value` which is not ideal. The current setup for values in
`engine` is quite complicated - there's 3 or 4 different enums that
represent values, and `serde::Deserialize` is currently the only way to
avoid dealing with that complexity. I did briefly try to write a
`Deserialize` impl instead of using `serde_json::Value` but that proved
to be more complicated than it was worth. If we ever refactor values in
engine we could probably revisit this.

Fixes GB-8014