cli: make publish idempotent (#2401)
Now the CLI will report when you publish a subgraph and neither the
subgraph schema nor the url changes (they are identical to what is
already published). In that case, no deployment will be created in the
schema registry. This is not an error, so the exit code remains 0.

The noise in the diff is from updating the platform's GraphQL schema (in
particular the api keys removal).

Companion platform PR: https://github.com/grafbase/api/pull/4269