chore: make engine use cynic-parser (#2402)
We decided a while ago to standardise on cynic-parser as it's faster & a
bit more ergonomic than `async-graphql-parser` (and therefore
`engine-parser`, which is just a fork of `async-graphql-parser`. So this
PR swaps the parser in `engine` to `cynic-parser`.

It's a fairly mechanical swap, the only complicating factor is
cynic-parser tracking Spans as byte offsets, whereas we need line &
column info for the `position` field in GraphQL errors. So I've added
some translation code for that.

Running our benchmarks this sees a 13-17% decrease in time in the
benchmarks without an operation cache. Although I wouldn't expect to see
similar results in production since the operation cache bypasses parsing
and waiting on subgraphs is likely to be much slower than parsing ever
was.