chore: stop using engine-value (#2413)
Now that we're not using engine-parser there's really no reason to keep
using engine-value. This swaps it for serde_json::Value, which is closer
to reality (since we're working with JSON) and has a slightly nicer API
(imo).
chore: stop using engine-value (#2413)

Now that we're not using engine-parser there's really no reason to keep
using engine-value. This swaps it for serde_json::Value, which is closer
to reality (since we're working with JSON) and has a slightly nicer API
(imo).
chore(deps): update dependencies-non-major