fix(engine): Fix requires-with-fragments
For requires with fragments we need to keep track of the object
definition id in the response object. So adding it. Need to rework some
things, but the test now passes.

Added a simple test in our own test suite as federation-audit tests
sometimes check a lot of things at once.