FederatedGraph: walk back TypeDefinition{Record,Id} (#2438)
There is a halfway done refactoring from Enum, Scalar, Object,
Interface, etc. to a unified TypeDefinition in the
graphql-federated-graph crate. The idea was to share the common fields
between all type definitions (name, description, directives) in a single
uniform struct, and store them all together. Many functions could then
work on type definitions directly without duplicating logic bbtween
different kinds of definitions, while still allowing specialized
handling for specific kinds of definitions. The expectation was that it
would simplify logic across graphql-composition and
graphql-federated-graph, and make graphql-federated-graph more readable
by reducing the amount of types in the public API and the amount of
fields in FederatedGraph. It would also make the `Definition` enum
unnecessary, since you could use a `TypeDefinitionId` instead.

After discussions with @Finistere, some experimentation and a lot of
thinking, I don't think this is going to pan out. A few reasons:

- When looking for a specific object, interface, etc., it's better to
search in a shorter Vec containing only the kinds of definitions we are
interested in.
- There are many places where we do want type definition specific logic.
- More precise ID types as a default are a better idea
- Separate type definition collections better match the model of
`engine_schema`.
- The code size savings are not that great.