Scope the CLI version to the grafbase crate
Until this commit, the CLI inherits its version from the workspace. So the workspace version is the CLI version. Other crates used to inherit it but that is no longer the case. The cli version shouldn't have precedence over versions for other crates that are decoupled from it, like the gateway or grafbase-composition.

So this commit scopes the version of the cli to the Cargo.toml of the grafbase crate.