chore: remove OperationCacheFactory trait (#2455)
To do query warming I need to reach into the operation cache to fetch
queries to prewarm. I wanted to do this in the runtime, but currently
the runtime doesn't have access to the operation cache, only an
OperationCacheFactory.

I can't see much reason for this factory. Things seem to work fine
without it, so I've removed it. If I've missed something please let me
know and I'll revert.