chore: rework engine & config updates in gateway (#2457)
I need to hook into the engine creation process in the gateway to add
query warming. Prior to this change that happened in three places:
inside the GraphUpdater when using a graph ref to fetch from the GDN & a
couple of places inside GraphFetchMethod::start when using a schema
file. This made it a bit awkward to add in query warming.

This refactors things a bit so that GraphUpdater &
GraphFetchMethod::start both just send a `GraphDefinition` on a channel,
where a new `EngineReloader` (name suggestions welcome) is listening and
will handle the update. There's also some logic to cancel a running
reload if another one comes in afterwards.

I also separated out config updates from graph updates in the
`reload_signal` used by federated dev - config updates now go over their
own `tokio::sync::watch` channel, rather than sometimes being passed as
an argument and sometimes being passed through a channel along with a
GraphDef.

Finally, I changed the engine watcher to not contain an `Option`. As far
as I could tell the case where this is `None` could only be ever be
observed by the websocket handler - we only ever set it to
`Some(engine)` and we have a `watcher.changed().await` before we start
up the main server. So it seems like a reasonable simplification.