fix: `grafbase-gateway` as default otel `service.name`
I've been using traces to test things locally lately by feeding things
into jaeger.  But if I forget to add a service name into my config my
traces appear in jaeger under `empty-service-name`.  This is because we
default to an empty string.

It seems better to have an actual default for this: I propose
`grafbase-gateway` but I'm open to other suggestions.