fix: add a default otel `service.name` (#2465)
I've been using traces to test things locally lately by feeding things
into jaeger. But if I forget to add a service name into my config my
traces appear in jaeger under `empty-service-name`. This is because we
default to an empty string for this config value.

It seems better to have an actual default for this: I propose
`grafbase-gateway` but I'm open to other suggestions. I'm assuming this
won't have an impact on our hosted otel stuff, but if someone more
familiar with it could confirm that'd be good.