feat: in memory prewarming on GDN updates (#2463)
The initial implementation of query warming: this will warm the entire
in-memory cache on schema or config updates in federated dev & the
gateway, if warming is enabled. A future PR will add more configuration
to this, at the moment it's just on/off.

One thing that may be of concern with this feature: a lot of the query
binding etc. code assumes it's occurring within the context of a
request. But warming is never occurring in a request. I've just set all
the associated values to null, but this seems like it could have some
consequences.

Fixes GB-8098