chore: extract what-rust-changed into its own action (#2477)
I changed the grafbase PR action to do partial builds a while back,
using what-rust-changed to decide what we need to build. This moves some
of the logic for that into a resuable action for what-rust-changed. I
initially wrote this because I was planning on using it again somewhere,
and intended to avoid repeating the pattern once again. But in the end I
didn't need to do that. This still seems like a reasonable cleanup to
make though.

Depends on https://github.com/grafbase/what-rust-changed/pull/7.