gateway: pass websocket connection_init payload to subgraphs
The gateway implements the [graphql-transport-ws protocol](https://github.com/graphql/graphql-over-http/blob/main/rfcs/GraphQLOverWebSocket.md). In that protocol, the first message on a websocket connection will be `connection_init`. That message contains an optional `payload` field that is a JSON object with arbitrary values. Typically, information that would go in HTTP headers for queries and mutations flows instead in that `connection_init` message's payload. This is because browsers can't send custom headers with the websocket HTTP upgrade request. So headers like `Authorization` are passed there instead.

Until this PR, the gateway expected a payload of the shape:

```ts
type ConnectionInit = {
  type: 'connection_init',
  payload: { headers: Map<string, string> }
}
```

It would then handle the contents of the `headers` key like headers, passing them through subgraph header rules, extracting JWTs from Authorization if so configured, etc. Eventually, these headers did land in the connection_init payload of requests made to subgraph, at the top level.

This is limiting, because both from the perspective of the client and the subgraphs, the payload is an arbitrary JSON object. We want subscriptions that work from a given client to a given subgraph to also work from the same client to the same subgraph with the Grafbase gateway in the middle.

This PR makes the gateway pass on the connection_init payload verbatim by default, with an option to disable it under:

```toml
[websockets]
forward_connection_init_payload = false
```

The gateway still attempts to extract headers for itself from the root of the connection_init payload and from its `headers` key if present.

This PR also contains new infrastructure to test websocket subscriptions in integration-tests, since the existing websocket tests were deleted with the CLI test suite.

closes GB-8244
closes GB-8262