fix(engine): Incorrect handling of missing query in request.
In our tests we were always sending an empty string and the previous
Request struct used to have `query: String`. But since the depreciation of
the previous engine we use `query: Option<String>` instead.

So now we weren't treating correctly a missing query field in the
request for APQ. Fixed that, added a few tests that send no `query`
field and put in common some code.

Fixes #2505