feat: schema file hot reload (#2521)
This will be enabled for everybody loading the schema file from a file.
Just start the gateway with the `-s` flag and it's running. Queries are
warmed accordingly as we do with the GDN updates.

Some notes on implementation. I decided on polling, due to notify
doesn't really work in certain scenarios:

- macOS with Docker doesn't have the events
- Any network filesystem does not have the events
- Possibly windows doesn't have the events
- And we are not sure about kubernetes config map filesystem...

We poll every five seconds, check the file metadata and its modified at.
If this is not available, we load the file contents, hash it to memory
and compare with that. This implementation should support the broadest
possible amount of use cases.