feat: very basic rest extension that can do a GET (#2533)
Adds a very simple REST extension, which should show how an extension
could look like. Notable things:

```rust
#[derive(grafbase_sdk::ResolverExtension)]
struct MyExtension;
```

This will generate the needed initialization code, and fail because the
struct needs to implement the `Extension` and `Resolver` traits. The
user implements these traits, the code compiles and should be ready to
use in the gateway.