gateway: release 0.28.2
**Features**

When the gateway receives a subscription request, it can use either server-sent events or websockets to communicate with the subgraph serving the subscription. This is independent from the protocol used by the client to connect to the gateway. Until now, the logic was that if a websocket_url was defined in the subgraph configuration, the gateway would use a websocket connection, otherwise default to server sent events. This is opaque, and not flexible enough.

This release introduces a new subgraph configuration option in grafbase.toml:

````
[subgraphs.my-subgraph]
subscriptions_protocol = "websocket"
````

The value must be either "websocket" or "server_sent_events". The gateway will follow the choice of protocol for the subgraph. If this option is omitted, the previous logic applies.